Time class moved
