+++ src/Model/Rent.java
@@ -5,6 +5,7 @@
  */
 package Model;
 
+import Time.Time;
 import java.util.Date;
 
 /**
