+++ src/Time/Time.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Model;
+package Time;
 
 import java.text.SimpleDateFormat;
 import java.time.LocalDate;
